% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ipw_rmst.R
\name{calc_ipw_rmst}
\alias{calc_ipw_rmst}
\title{IPW-weighting Model}
\usage{
calc_ipw_rmst(data, A, time, status, covariates, tau)
}
\arguments{
\item{data:}{survival data}

\item{A:}{treatment}

\item{time:}{observed survival time}

\item{status:}{survival status}

\item{covariates:}{confounders}

\item{tau:}{time point of interest}
}
\value{
IPW-RMST
}
\description{
IPW-weighting Model
}
