% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OAL-Cox.R
\name{OAL_Cox}
\alias{OAL_Cox}
\title{Outcome-adaptive Lasso for Cox Proportional Hazard Model}
\usage{
OAL_Cox(
  data,
  A,
  time,
  status,
  covariates,
  lambda_vec,
  gamma_convergence_factor,
  tau
)
}
\arguments{
\item{data:}{survival data}

\item{A:}{treatment}

\item{time:}{observed survival time}

\item{status:}{survival status}

\item{covariates:}{confounders}

\item{lambda_vec:}{regularization parameter for OAL-Cox model}

\item{gamma_convergence_factor:}{converge factor}

\item{tau:}{time point of interest}
}
\value{
RASCE of OAL-Cox model, the optimal lambda_n and propensity score coefficients corresponding to this value
}
\description{
Outcome-adaptive Lasso for Cox Proportional Hazard Model
}
